requirements.txt
streamlit
langchain-google-genai
langchain-community
langgraph
tavily-python
python-dotenv
PyPDF2
google-api-python-client
google-auth